+++ src/functions.h
@@ -1,6 +1,6 @@
 #ifndef FUNCTIONS_H
 #define FUNCTIONS_H
 
-float nr_sqrt(float n);
+double nr_sqrt(double n);
 
 #endif
